Update table
git-svn-id: svn://at-lab/Labview@3754 d1c95485-6871-e240-9df0-63fa95a8ea69

+++ Acquisition/Database/timestamper.sql
@@ -5,7 +5,7 @@ USE sessions;
 
 CREATE TABLE sessions (
 	setup TINYINT UNSIGNED NOT NULL,
-	session_start_time BIGINT NOT NULL,
+	session_start_time DOUBLE NOT NULL,
 	subject VARCHAR(45) NOT NULL,
 	experimenter VARCHAR(45) NOT NULL,
 	session_path LONGTEXT NOT NULL,
@@ -14,10 +14,10 @@ CREATE TABLE sessions (
 
 CREATE TABLE session_timestamps (
 	setup TINYINT UNSIGNED NOT NULL,
-	session_start_time BIGINT NOT NULL,
+	session_start_time DOUBLE NOT NULL,
 	channel INT NOT NULL,
 	count INT UNSIGNED NOT NULL,
-	time BIGINT NOT NULL,
+	time DOUBLE NOT NULL,
 	FOREIGN KEY (setup, session_start_time) REFERENCES sessions(setup,session_start_time) ON DELETE RESTRICT ON UPDATE RESTRICT,
 	PRIMARY KEY (setup, session_start_time, channel, count)
 );
@@ -25,8 +25,8 @@ CREATE TABLE session_timestamps (
 # Table to record all the electrophysiology recordings and link them to sessions
 CREATE TABLE ephys (
 	setup TINYINT UNSIGNED NOT NULL,
-	session_start_time BIGINT NOT NULL,
-	ephys_start_time BIGINT NOT NULL,
+	session_start_time DOUBLE NOT NULL,
+	ephys_start_time DOUBLE NOT NULL,
 	ephys_path LONGTEXT NOT NULL,
 	FOREIGN KEY (setup, session_start_time) REFERENCES sessions(setup,session_start_time) ON DELETE RESTRICT ON UPDATE RESTRICT,
 	PRIMARY KEY (setup, session_start_time, ephys_start_time)
@@ -35,8 +35,8 @@ CREATE TABLE ephys (
 # Table of stimulation sessions
 CREATE TABLE stimulation (
 	setup TINYINT UNSIGNED NOT NULL,
-	session_start_time BIGINT NOT NULL,
-	stim_start_time BIGINT NOT NULL,
+	session_start_time DOUBLE NOT NULL,
+	stim_start_time DOUBLE NOT NULL,
 	stim_path LONGTEXT NOT NULL,
 	exp_type VARCHAR(45) NOT NULL,
 	total_trials INT,
@@ -49,9 +49,9 @@ CREATE TABLE stimulation (
 # Table of stimulation sessions
 CREATE TABLE tber_pulses(
 	setup TINYINT UNSIGNED NOT NULL,
-	session_start_time BIGINT NOT NULL,
-	stim_start_time BIGINT NOT NULL,
-	time BIGINT NOT NULL,
+	session_start_time DOUBLE NOT NULL,
+	stim_start_time DOUBLE NOT NULL,
+	time DOUBLE NOT NULL,
 	FOREIGN KEY (setup, session_start_time, stim_start_time) REFERENCES stimulation(setup, session_start_time, stim_start_time) ON DELETE RESTRICT ON UPDATE RESTRICT,
 	PRIMARY KEY (setup, session_start_time, stim_start_time,time)
 );
@@ -59,9 +59,9 @@ CREATE TABLE tber_pulses(
 # Table for all the traces
 CREATE TABLE behavior_traces (
 	setup TINYINT UNSIGNED NOT NULL,
-	session_start_time BIGINT NOT NULL,
-	stim_start_time BIGINT NOT NULL,
-	beh_start_time BIGINT NOT NULL,
+	session_start_time DOUBLE NOT NULL,
+	stim_start_time DOUBLE NOT NULL,
+	beh_start_time DOUBLE NOT NULL,
 	beh_path LONGTEXT NOT NULL,
 	beh_traces_type VARCHAR(45) NOT NULL,
 	FOREIGN KEY (setup, session_start_time, stim_start_time) REFERENCES stimulation(setup,session_start_time,stim_start_time) ON DELETE RESTRICT ON UPDATE RESTRICT,
